Trabalho 6/sistema-carga-ia/consumer_team/requirements.txt
pika==1.2.0
Pillow==9.4.0
python-dotenv==0.21.0
tensorflow==2.12.0
opencv-python-headless==4.7.0.72
numpy==1.24.2
matplotlib==3.7.1
scikit-learn==1.2.2
